fix: Rename end-to-end NG tests to not contain NG
Connects to #4450.
